pyproject.toml
[project]
name = "hackaton"
version = "0.1.0"
description = "Telegram-бот с поддержкой AI-моделей через OpenAI-совместимое API"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[rate-limiter]==21.8",
    "python-dotenv>=1.0",
    "openai>=1.52",
    "httpx>=0.27",
    "tenacity>=9.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "ruff>=0.7",
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"
